moves TranslationMapFetcher fucntionailty to TranslationMapGenerator
