+++ umich_catalog_indexing/lib/jobs/utilities.rb
@@ -1,6 +1,5 @@
 require "jobs/utilities/alma_file_processor"
 require "jobs/utilities/zephir_file"
-require "jobs/utilities/translation_map_fetcher"
 
 module Jobs
   module Utilities
